requirements.txt
flask==2.3.3
flask-cors==4.0.0
pillow==10.4.0
rembg==2.0.57
numpy==1.26.4
scikit-image==0.22.0
werkzeug==3.0.3